requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
binaryornot==0.4.4
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
eventlet==0.37.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.1
